solve conflict of the user model
