+++ firefly/templates/base.html
@@ -1,6 +1,6 @@
 -py
   from flask import url_for
-  from flask.ext.login import current_user
+  from flask_login import current_user
 
 doctype html
 html
